Refactor db pool handling and passport code
