+++ src/db/pool.js
@@ -1,0 +1,33 @@
+/*
+ * Copyright (c) 2024, WSO2 LLC. (http://www.wso2.com) All Rights Reserved.
+ *
+ * WSO2 LLC. licenses this file to you under the Apache License,
+ * Version 2.0 (the "License"); you may not use this file except
+ * in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing,
+ * software distributed under the License is distributed on an
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ * KIND, either express or implied. See the License for the
+ * specific language governing permissions and limitations
+ * under the License.
+ */
+
+const { Pool } = require('pg');
+const { config, secrets: secretConf } = require('../config/configLoader');
+
+const poolConfig = {
+    user: config.db.username,
+    host: config.db.host,
+    database: config.db.database,
+    password: secretConf.dbSecret,
+    port: config.db.port,
+    ...(config.advanced.dbSslDialectOption && { ssl: { require: true, rejectUnauthorized: false } }),
+};
+
+const pool = new Pool(poolConfig);
+
+module.exports = pool;
